refactor: extract RenderFramePersistenceStore
